* Only blow away the database on test pass if we set one up in the first place
git-svn-id: svn+ssh://svn.bestpractical.com/svn/jifty.org/jifty/trunk@911 e84bef0a-9b06-0410-84ba-c4c9edb13aeb

+++ lib/Jifty/Test.pm
@@ -182,12 +182,14 @@ END {
         unlink mailbox();
 
         # Remove testing db
-        Jifty->handle->disconnect() if (Jifty->handle);
+        if (Jifty->handle) {
+        Jifty->handle->disconnect();
         Log::Log4perl->get_logger("SchemaTool")->less_logging(3);
         my $schema = Jifty::Script::Schema->new;
         $schema->{drop_database} = 1;
         $schema->run;
         Log::Log4perl->get_logger("SchemaTool")->more_logging(3);
+        }
     }
 }
